Ограничены права на работу модуля "block_stickerpack", фильтры вынесены в отдельный файл
